backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
aleph-sdk-python==1.1.0
eth-account==0.13.4
pydantic==2.10.3
aiofiles==24.1.0
httpx==0.28.1
python-dotenv==1.0.1
